split files & add consts
